Isolate all WebAuth files from non iOS targets
